requirements.txt
streamlit==1.28.0
jira==3.3.1
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.25.2
pandas==2.1.1
groq==0.0.8
